+++ features/config/sync-strategy/update.feature
@@ -1,7 +1,7 @@
 Feature: configure the sync-strategy
 
   Scenario Outline:
-    When I run "git-town sync-strategy <VALUE>"
+    When I run "git-town config sync-strategy <VALUE>"
     Then setting "sync-strategy" is now "<VALUE>"
 
     Examples:
